+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## dev
+
+
+
 ## 1.0.4 (2024-10-07)
 
 * Added setting `ENVELOPE_USER_CHANNEL_SEND_SUBSCRIBE` to send a subscribe message to consumer rather than just adding